Refactor campaign analytics processing to use batch fetching and modularize utility functions
